requirements.txt
Package            Version
------------------ --------
asttokens          2.0.8
backcall           0.2.0
brotlipy           0.7.0
certifi            2023.5.7
cffi               1.15.1
charset-normalizer 2.0.4
colorama           0.4.6
cryptography       39.0.1
debugpy            1.6.3
decorator          5.1.1
docopt             0.6.2
executing          1.1.0
gitdb              4.0.10
GitPython          3.1.32
idna               3.4
ipykernel          6.16.0
ipython            8.5.0
jedi               0.18.1
jsonpickle         3.0.1
jupyter_client     7.3.5
jupyter-core       4.11.1
matplotlib-inline  0.1.6
mkl-fft            1.3.6
mkl-random         1.2.2
mkl-service        2.4.0
munch              2.5.0
nest-asyncio       1.5.5
numpy              1.24.3
packaging          23.1
pandas             2.0.3
parso              0.8.3
pickleshare        0.7.5
Pillow             9.4.0
pip                23.1.2
prompt-toolkit     3.0.31
protobuf           3.20.0
psutil             5.9.2
pure-eval          0.2.2
py-cpuinfo         9.0.0
pycparser          2.21
Pygments           2.13.0
pyOpenSSL          23.0.0
pyparsing          3.0.9
PySocks            1.7.1
python-dateutil    2.8.2
pytz               2023.3
PyYAML             6.0.1
pyzmq              24.0.1
requests           2.29.0
sacred             0.8.4
scipy              1.10.1
setuptools         67.8.0
six                1.16.0
smmap              5.0.0
stack-data         0.5.1
tensorboard-logger 0.1.0
torch              1.11.0
torchaudio         0.11.0
torchvision        0.12.0
tornado            6.2
tqdm               4.65.0
traitlets          5.4.0
typing_extensions  4.6.3
tzdata             2023.3
urllib3            1.26.16
wcwidth            0.2.5
wheel              0.38.4
wrapt              1.15.0
